+++ piecrust/CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Apply charging mechanism for host queries [#359]
+- Add `HostQuery::execute` and `HostQuery::deserialize_and_price` [#359]
+
 ## Changed
 
+- Drop `Fn(&mut [u8], u32) -> u32` bound for `HostQuery` [#359]
 - Make storage instructions cost 4 gas per byte [#359]
 - Upgrade `dusk-wasmtime` to version `21.0.0-alpha` [#359]